# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Neighbors",
        "version": "release 1.0"
    },
    "paths": {
...
        "/api/v1/location": {
            "get": {
                "tags": [
                    "Barrios"
                ],
                "summary": "Listar los barrios",
                "description": "Listar los barrios (con Token)",
                "operationId": "LocationsGet",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ok"
                                        },
                                        "obj": {
                                            "type": "string",
                                            "example": "array()"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Barrios"
                ],
                "summary": "Modificar un Barrio",
                "description": "Modificar un Barrio (con Token)",
                "operationId": "LocationsUpdate",
                "requestBody": {
                    "description": "Datos del Barrio",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "int",
                                        "format": "number",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "La arboleda"
                                    },
                                    "altitude": {
                                        "type": "double",
                                        "format": "number",
                                        "example": 13.22132
                                    },
                                    "latitude": {
                                        "type": "double",
                                        "format": "number",
                                        "example": 13.22132
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Location successfully updated"
                                        },
                                        "objLocation": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "La arboleda"
                                                },
                                                "altitude": {
                                                    "type": "double",
                                                    "format": "number",
                                                    "example": 13.22132
                                                },
                                                "latitude": {
                                                    "type": "double",
                                                    "format": "number",
                                                    "example": 13.22132
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2021-01-18T15:31:11.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2021-01-18T15:31:11.000000Z"
                                                },
                                                "id": {
                                                    "type": "int",
                                                    "format": "number",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "The id field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Barrios"
                ],
                "summary": "Crear un Barrio",
                "description": "Crear un Barrio (con Token)",
                "operationId": "LocationsCreate",
                "requestBody": {
                    "description": "Datos del Barrio",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "La arboleda"
                                    },
                                    "altitude": {
                                        "type": "double",
                                        "format": "number",
                                        "example": 13.22132
                                    },
                                    "latitude": {
                                        "type": "double",
                                        "format": "number",
                                        "example": 13.22132
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Location successfully registered"
                                        },
                                        "objLocation": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "La arboleda"
                                                },
                                                "altitude": {
                                                    "type": "double",
                                                    "format": "number",
                                                    "example": 13.22132
                                                },
                                                "latitude": {
                                                    "type": "double",
                                                    "format": "number",
                                                    "example": 13.22132
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2021-01-18T15:31:11.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2021-01-18T15:31:11.000000Z"
                                                },
                                                "id": {
                                                    "type": "int",
                                                    "format": "number",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/location/state": {
            "get": {
                "tags": [
                    "Provincias"
                ],
                "summary": "Listar las provincias",
                "description": "Listar las provincias (con Token)",
                "operationId": "LocationsGet",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ok"
                                        },
                                        "obj": {
                                            "type": "string",
                                            "example": "array()"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/v1/location/district/{state_id}": {
            "get": {
                "tags": [
                    "Localidades"
                ],
                "summary": "Listar de localidades",
                "description": "Listar de localidades (con Token)",
                "operationId": "LocationsGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del mensaje",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ok"
                                        },
                                        "obj": {
                                            "type": "string",
                                            "example": "array()"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Header autorization toker JWT",
                "name": "Token Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}